Cleanup repository root
RELEASE_NOTES=n/a
