+++ .gitignore
@@ -1,4 +1,6 @@
 gopass
+gopass-*-amd64
+gopass-full
 !pkg/gopass/
 coverage.out
 coverage-all.*
