+++ main_test.go
@@ -2,12 +2,25 @@ package main
 
 import (
 	"bytes"
+	"context"
+	"flag"
 	"fmt"
+	"os"
 	"runtime"
 	"testing"
 
+	"github.com/atotto/clipboard"
 	"github.com/blang/semver"
+	"github.com/fatih/color"
+	"github.com/gopasspw/gopass/internal/action"
+	"github.com/gopasspw/gopass/internal/backend"
+	"github.com/gopasspw/gopass/internal/backend/crypto/gpg"
+	"github.com/gopasspw/gopass/internal/config"
+	"github.com/gopasspw/gopass/internal/gptest"
+	"github.com/gopasspw/gopass/internal/out"
+	"github.com/gopasspw/gopass/pkg/ctxutil"
 	"github.com/stretchr/testify/assert"
+	"github.com/urfave/cli/v2"
 )
 
 func TestVersionPrinter(t *testing.T) {
@@ -27,4 +40,130 @@ func TestGetVersion(t *testing.T) {
 	if getVersion().LT(semver.Version{Major: 1, Minor: 9}) {
 		t.Errorf("invalid version")
 	}
+}
+
+func TestSetupApp(t *testing.T) {
+	ctx := context.Background()
+	_, app := setupApp(ctx, semver.Version{})
+	assert.NotNil(t, app)
+}
+
+// commandsWithError is a list of commands that return an error when
+// invoked without arguments
+var commandsWithError = map[string]struct{}{
+	".audit":             {},
+	".audit.hibp":        {},
+	".binary.cat":        {},
+	".binary.copy":       {},
+	".binary.move":       {},
+	".binary.sum":        {},
+	".clone":             {},
+	".copy":              {},
+	".create":            {},
+	".delete":            {},
+	".edit":              {},
+	".env":               {},
+	".find":              {},
+	".generate":          {},
+	".git.push":          {},
+	".git.remote.add":    {},
+	".git.remote.remove": {},
+	".grep":              {},
+	".history":           {},
+	".init":              {},
+	".insert":            {},
+	".mounts.add":        {},
+	".mounts.remove":     {},
+	".move":              {},
+	".otp":               {},
+	".recipients.add":    {},
+	".recipients.remove": {},
+	".setup":             {},
+	".show":              {},
+	".templates.edit":    {},
+	".templates.remove":  {},
+	".templates.show":    {},
+	".unclip":            {},
+}
+
+func TestGetCommands(t *testing.T) {
+	u := gptest.NewUnitTester(t)
+	defer u.Remove()
+
+	buf := &bytes.Buffer{}
+	out.Stdout = buf
+	color.NoColor = true
+	defer func() {
+		out.Stdout = os.Stdout
+	}()
+
+	cfg := config.New()
+	cfg.Root.Path = backend.FromPath(u.StoreDir(""))
+
+	clipboard.Unsupported = true
+
+	ctx := context.Background()
+	ctx = ctxutil.WithAlwaysYes(ctx, true)
+	ctx = ctxutil.WithInteractive(ctx, false)
+	ctx = ctxutil.WithTerminal(ctx, false)
+	ctx = out.WithHidden(ctx, true)
+	ctx = backend.WithRCSBackendString(ctx, "noop")
+	ctx = backend.WithCryptoBackendString(ctx, "plain")
+
+	act, err := action.New(ctx, cfg, semver.Version{})
+	assert.NoError(t, err)
+
+	app := cli.NewApp()
+	fs := flag.NewFlagSet("default", flag.ContinueOnError)
+	c := cli.NewContext(app, fs, nil)
+	c.Context = ctx
+
+	commands := getCommands(ctx, act, app)
+	assert.Equal(t, 32, len(commands))
+
+	prefix := ""
+	testCommands(t, c, commands, prefix)
+}
+
+func testCommands(t *testing.T, c *cli.Context, commands []*cli.Command, prefix string) {
+	for _, cmd := range commands {
+		if cmd.Name == "update" {
+			continue
+		}
+		if len(cmd.Subcommands) > 0 {
+			testCommands(t, c, cmd.Subcommands, prefix+"."+cmd.Name)
+		}
+		if cmd.Before != nil {
+			if err := cmd.Before(c); err != nil {
+				continue
+			}
+		}
+		if cmd.BashComplete != nil {
+			cmd.BashComplete(c)
+		}
+		if cmd.Action != nil {
+			fullName := prefix + "." + cmd.Name
+			if _, found := commandsWithError[fullName]; found {
+				assert.Error(t, cmd.Action(c), fullName)
+				continue
+			}
+			assert.NoError(t, cmd.Action(c), fullName)
+		}
+	}
+}
+
+func TestInitContext(t *testing.T) {
+	ctx := context.Background()
+	cfg := config.New()
+
+	ctx = initContext(ctx, cfg)
+	assert.Equal(t, true, gpg.IsAlwaysTrust(ctx))
+
+	assert.NoError(t, os.Setenv("GOPASS_DEBUG", "true"))
+	ctx = initContext(ctx, cfg)
+	assert.Equal(t, true, ctxutil.IsDebug(ctx))
+
+	assert.NoError(t, os.Setenv("GOPASS_NOCOLOR", "true"))
+	ctx = initContext(ctx, cfg)
+	assert.Equal(t, false, ctxutil.IsColor(ctx))
 }
